+++ website/utils/snippets.py
@@ -6,6 +6,8 @@ from collections import namedtuple
 from _sha1 import sha1
 
 from django.conf import settings
+from django.core import mail
+from django.template import loader
 from django.template.defaultfilters import urlencode
 from django.templatetags.static import static
 from django.utils import timezone, dateparse
@@ -182,3 +184,20 @@ def overlaps(check, others, can_equal=True):
             return True
 
     return False
+
+
+def _send_email(to: str, subject: str, body_template: str, context: dict) -> None:
+    """
+    Easily send an email with the right subject and a body template
+
+    :param to: where should the email go?
+    :param subject: what is the email about?
+    :param body_template: what is the content of the email?
+    :param context: add some context to the body
+    """
+    mail.EmailMessage(
+        "[THALIA] {}".format(subject),
+        loader.render_to_string(body_template, context),
+        settings.DEFAULT_FROM_EMAIL,
+        [to],
+    ).send()
